backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
SQLAlchemy==2.0.23
mysqlclient==2.2.0
pymysql==1.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
httpx==0.25.1
python-multipart==0.0.6
pydantic-settings==2.1.0
requests==2.31.0
